+++ crates/reqmd_core/src/support/fixtures.rs
@@ -10,6 +10,15 @@ pub fn sample_env_var_expansion_request_file() -> File {
     File::load(include_str!("./fixtures/sample-env-expansion.md"), None).unwrap()
 }
 
+#[rstest::fixture]
+pub fn sample_server_from_hostname_request_file() -> File {
+    File::load(
+        include_str!("./fixtures/sample-server-from-hostname.md"),
+        None,
+    )
+    .unwrap()
+}
+
 #[cfg(feature = "yaml-as-json")]
 #[rstest::fixture]
 pub fn sample_yaml_as_json_request_file() -> File {